Update to Web3 1.0-beta51. Nothing works.
This is some crap actually. It almost exactly a year since I touched this lib. Pause with
a promise to push it to 1.0 when Web3 is 1.0. They added 25 versions of betas
since then and it is the absolutely worst experience I have had with something
from TypeScript. A feeling is that someone on a team is sabotaging TS and from
the camp 'we do not need types because they will reveal how many bugs we have'...
Apart from broken types every release breaks somethings, from ABI encoding
and ongoing mess with fuck knows what version of BigNumber/BN is the only
right one to simple confirmation of contract deployment transaction.